maintain stop time on recompile
