+++ pkg/cli/commands_test.go
@@ -979,94 +979,7 @@ This workflow uses an include.
 	// previewOrphanedIncludes and the flag handling is straightforward.
 }
 
-// TestExtractStopTimeFromLockFile tests the extractStopTimeFromLockFile function
-func TestExtractStopTimeFromLockFile(t *testing.T) {
-	// Create temporary directory for test files
-	tmpDir := t.TempDir()
 
-	tests := []struct {
-		name         string
-		content      string
-		expectedTime string
-	}{
-		{
-			name: "valid stop-time in lock file",
-			content: `# This file was automatically generated by gh-aw
-name: "Test Workflow"
-jobs:
-  test:
-    steps:
-      - name: Safety checks
-        run: |
-          STOP_TIME="2025-12-31 23:59:59"
-          echo "Checking stop-time limit: $STOP_TIME"`,
-			expectedTime: "2025-12-31 23:59:59",
-		},
-		{
-			name: "no stop-time in lock file",
-			content: `# This file was automatically generated by gh-aw
-name: "Test Workflow"
-jobs:
-  test:
-    steps:
-      - name: Run tests
-        run: echo "No stop time here"`,
-			expectedTime: "",
-		},
-		{
-			name: "malformed stop-time line",
-			content: `# This file was automatically generated by gh-aw
-name: "Test Workflow"
-jobs:
-  test:
-    steps:
-      - name: Safety checks
-        run: |
-          STOP_TIME=malformed-no-quotes
-          echo "Invalid format"`,
-			expectedTime: "",
-		},
-		{
-			name: "multiple stop-time lines (should get first)",
-			content: `# This file was automatically generated by gh-aw
-name: "Test Workflow"
-jobs:
-  test:
-    steps:
-      - name: Safety checks
-        run: |
-          STOP_TIME="2025-06-01 12:00:00"
-          echo "Checking stop-time limit: $STOP_TIME"
-          STOP_TIME="2025-07-01 12:00:00"`,
-			expectedTime: "2025-06-01 12:00:00",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			// Create test lock file
-			lockFile := filepath.Join(tmpDir, tt.name+".lock.yml")
-			err := os.WriteFile(lockFile, []byte(tt.content), 0644)
-			if err != nil {
-				t.Fatalf("Failed to create test file: %v", err)
-			}
-
-			// Test extraction
-			result := extractStopTimeFromLockFile(lockFile)
-			if result != tt.expectedTime {
-				t.Errorf("extractStopTimeFromLockFile() = %q, want %q", result, tt.expectedTime)
-			}
-		})
-	}
-
-	// Test non-existent file
-	t.Run("non-existent file", func(t *testing.T) {
-		result := extractStopTimeFromLockFile("/non/existent/file.lock.yml")
-		if result != "" {
-			t.Errorf("extractStopTimeFromLockFile() for non-existent file = %q, want empty string", result)
-		}
-	})
-}
 
 // TestCalculateTimeRemaining tests the calculateTimeRemaining function
 func TestCalculateTimeRemaining(t *testing.T) {
